Move format checkers to dockstore-common
