+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/RunExecution.java
@@ -24,11 +24,6 @@ import io.dockstore.webservice.core.metrics.constraints.ISO8601ExecutionTime;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.media.Schema.RequiredMode;
 import jakarta.validation.constraints.NotNull;
-import java.time.Duration;
-import java.time.format.DateTimeParseException;
-import java.util.Optional;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * This is an object to encapsulate workflow run execution metrics data in an entity. Does not need to be stored in the database.
@@ -36,7 +31,6 @@ import org.slf4j.LoggerFactory;
 @Schema(name = "RunExecution", description = "Metrics of a workflow execution on a platform", allOf = Execution.class)
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public class RunExecution extends Execution {
-    private static final Logger LOG = LoggerFactory.getLogger(RunExecution.class);
 
     @NotNull
     @JsonProperty(required = true)
@@ -93,19 +87,5 @@ public class RunExecution extends Execution {
 
     public void setCpuRequirements(Integer cpuRequirements) {
         this.cpuRequirements = cpuRequirements;
-    }
-
-    /**
-     * Check that the execution time is in ISO-8601 format by parsing it into a Duration.
-     * @param executionTime ISO 8601 execution time
-     * @return Duration parsed from the ISO 8601 execution time
-     */
-    public static Optional<Duration> checkExecutionTimeISO8601Format(String executionTime) {
-        try {
-            return Optional.of(Duration.parse(executionTime));
-        } catch (DateTimeParseException e) {
-            LOG.warn("Execution time {} is not in ISO 8601 format and could not be parsed to a Duration", executionTime, e);
-            return Optional.empty();
-        }
     }
 }
