+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/constraints/ISO8601ExecutionDateValidator.java
@@ -17,7 +17,7 @@
 
 package io.dockstore.webservice.core.metrics.constraints;
 
-import io.dockstore.webservice.core.metrics.ValidationExecution;
+import io.dockstore.common.metrics.FormatCheckHelper;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
@@ -31,6 +31,6 @@ public class ISO8601ExecutionDateValidator implements ConstraintValidator<ISO860
         if (executionTime == null) {
             return false;
         }
-        return ValidationExecution.checkExecutionDateISO8601Format(executionTime).isPresent();
+        return FormatCheckHelper.checkExecutionDateISO8601Format(executionTime).isPresent();
     }
 }
