+++ dockstore-webservice/src/main/java/io/dockstore/webservice/core/metrics/constraints/ISO8601ExecutionTimeValidator.java
@@ -17,7 +17,8 @@
 
 package io.dockstore.webservice.core.metrics.constraints;
 
-import io.dockstore.webservice.core.metrics.RunExecution;
+import static io.dockstore.common.metrics.FormatCheckHelper.checkExecutionTimeISO8601Format;
+
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
@@ -32,6 +33,6 @@ public class ISO8601ExecutionTimeValidator implements ConstraintValidator<ISO860
         if (executionTime == null) {
             return true;
         }
-        return RunExecution.checkExecutionTimeISO8601Format(executionTime).isPresent();
+        return checkExecutionTimeISO8601Format(executionTime).isPresent();
     }
 }
